+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 * Minors:
   * Add docker configuration for development purposes.
   * Configured CS Fixer.
+  * Rename class `\Aeliot\Bundle\TransMaintain\Model\CsvReader` to `\Aeliot\Bundle\TransMaintain\Model\CSV`.
   * Refactored gluing of yaml tree to single level array.
 * Bug fixes:
   * Fix registering of missed translation.